refactor: organize files and split impls
